requirements.txt
openai>=1.12.0
chromadb>=0.4.22
pgvector>=0.2.4
psycopg2-binary>=2.9.9
weaviate-client>=4.4.0
pinecone-client>=3.0.0
sentence-transformers>=2.3.0
torch>=2.1.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
tqdm>=4.66.0
datasets>=2.17.0
